fix: move tests into root/test/ directory
