--- src/test/esm/default-esm-export.mjs
@@ -1,3 +1,0 @@
-import passport from '../../../dist/index.js'
-
-passport.initialize()
