--- src/test/esm/named-esm-export.mjs
@@ -1,3 +1,0 @@
-import { Strategy } from '../../../dist/index.js'
-
-class MS extends Strategy {}
