+++ test/esm/default-esm-export.mjs
@@ -1,0 +1,3 @@
+import passport from '../../dist/src/index.js'
+
+passport.initialize()
