+++ test/esm/named-esm-export.mjs
@@ -1,0 +1,3 @@
+import { Strategy } from '../../dist/src/index.js'
+
+class MS extends Strategy {}
