+++ test/helpers.ts
@@ -3,15 +3,15 @@ import { join } from 'node:path'
 import fastify, { FastifyInstance } from 'fastify'
 import fastifySecureSession, { SecureSessionPluginOptions } from '@fastify/secure-session'
 import fastifyCookie from '@fastify/cookie'
-import Authenticator, { AuthenticatorOptions } from '../Authenticator'
-import { Strategy } from '../strategies'
+import Authenticator, { AuthenticatorOptions } from '../src/Authenticator'
+import { Strategy } from '../src/strategies'
 import { InjectOptions, Response as LightMyRequestResponse } from 'light-my-request'
 import parseCookies from 'set-cookie-parser'
 import { IncomingMessage } from 'node:http'
 import { FastifyRegisterOptions } from 'fastify/types/register'
 import { fastifySession, FastifySessionOptions } from '@fastify/session'
 
-const SecretKey = fs.readFileSync(join(__dirname, '../../src/test', 'secure.key'))
+const SecretKey = fs.readFileSync(join(__dirname, '../../test', 'secure.key'))
 
 let counter = 0
 export const generateTestUser = () => ({ name: 'test', id: String(counter++) })
